Refactor ast_to_ast transformations and create an interface for the preprocessor
