+++ compiler/ast_util.py
@@ -27,3 +27,27 @@ class PreprocessedAST:
 class UnparsedScript:
     def __init__(self, text):
         self.text = text
+
+
+##
+## Pattern matching for the AST
+##
+
+def check_if_ast_is_supported(construct, arguments, **kwargs):
+    return
+
+def ast_match_untyped(untyped_ast_object, cases, *args):
+    ## TODO: This should construct the complete AstNode object (not just the surface level)
+    ## TODO: Remove this and then at some point make real proper use of the AstNode
+    ast_node = AstNode(untyped_ast_object)
+    if ast_node.construct is AstNodeConstructor.PIPE:
+        ast_node.check(children_count = lambda : len(ast_node.items) >= 2)
+    return ast_match(ast_node, cases, *args)
+
+def ast_match(ast_node, cases, *args):
+    ## TODO: Remove that once `ast_match_untyped` is fixed to
+    ##       construct the whole AstNode object.
+    if(not isinstance(ast_node, AstNode)):
+        return ast_match_untyped(ast_node, cases, *args)
+
+    return cases[ast_node.construct.value](*args)(ast_node)
